refactor(telemetry): move the telemetry primitives into the light package
`attributes.py`, `context.py` and `utils/token_extraction.py` sat in `rhesis-sdk`
but import nothing heavier than stdlib and `rhesis.telemetry.schemas`, which is
already in the light package. Anything wanting the AI semantic conventions had to
take the whole SDK with them: 198 resolved packages, including torch, 15 nvidia-*
CUDA wheels, transformers and a `deepeval==3.7.0` pin — plus a Python 3.12 floor.

Moved to `rhesis.telemetry.*`. `rhesis[telemetry]` now resolves to 22 packages on
Python 3.10, with nothing heavy in it, and carries every symbol a framework
integration needs.

The 71 existing import sites are untouched: all three old paths are now shim
modules re-exporting the same objects, so `rhesis.sdk.telemetry.context` and
`rhesis.telemetry.context` read and write the same ContextVars rather than two
sets of them. Verified in both directions, and `apps/backend` imports unchanged.
`packages/rhesis` needed no new dependencies.

`make test` in sdk/: 2354 passed, 16 skipped.
