+++ packages/rhesis/src/rhesis/telemetry/__init__.py
@@ -1,5 +1,29 @@
 """Lightweight OpenTelemetry telemetry for the Rhesis platform."""
 
+from rhesis.telemetry.attributes import (
+    MAX_CONTENT_LENGTH,
+    AIAttributes,
+    AIEvents,
+    create_llm_attributes,
+    create_tool_attributes,
+    validate_span_name,
+)
+from rhesis.telemetry.context import (
+    get_conversation_id,
+    get_conversation_mapped_input,
+    get_conversation_trace_id,
+    get_root_trace_id,
+    get_test_execution_context,
+    is_llm_observation_active,
+    is_tracing_disabled,
+    set_conversation_id,
+    set_conversation_mapped_input,
+    set_conversation_trace_id,
+    set_llm_observation_active,
+    set_root_trace_id,
+    set_test_execution_context,
+    set_tracing_disabled,
+)
 from rhesis.telemetry.exporter import RhesisOTLPExporter
 from rhesis.telemetry.provider import get_tracer_provider, shutdown_tracer_provider
 from rhesis.telemetry.schemas import (
@@ -10,6 +34,7 @@ from rhesis.telemetry.schemas import (
     SpanLink,
     StatusCode,
 )
+from rhesis.telemetry.token_extraction import extract_token_usage, get_first_value
 
 __all__ = [
     "RhesisOTLPExporter",
@@ -21,4 +46,29 @@ __all__ = [
     "SpanKind",
     "SpanLink",
     "StatusCode",
+    # Semantic conventions
+    "MAX_CONTENT_LENGTH",
+    "AIAttributes",
+    "AIEvents",
+    "create_llm_attributes",
+    "create_tool_attributes",
+    "validate_span_name",
+    # Context
+    "get_conversation_id",
+    "get_conversation_mapped_input",
+    "get_conversation_trace_id",
+    "get_root_trace_id",
+    "get_test_execution_context",
+    "is_llm_observation_active",
+    "is_tracing_disabled",
+    "set_conversation_id",
+    "set_conversation_mapped_input",
+    "set_conversation_trace_id",
+    "set_llm_observation_active",
+    "set_root_trace_id",
+    "set_test_execution_context",
+    "set_tracing_disabled",
+    # Token usage
+    "extract_token_usage",
+    "get_first_value",
 ]
